+++ plugins/org.preesm.codegen.xtend/src/org/preesm/codegen/xtend/spider/visitor/SpiderCodegen.java
@@ -70,6 +70,7 @@ import org.preesm.model.scenario.Timing;
 import org.preesm.model.scenario.papi.PapiEvent;
 import org.preesm.model.scenario.papi.PapifyConfigActor;
 import org.preesm.model.scenario.papi.PapifyConfigManager;
+import org.preesm.model.scenario.papi.PapifyConfigPE;
 import org.preesm.model.slam.ComponentInstance;
 import org.preesm.model.slam.Design;
 import org.preesm.model.slam.component.Component;
@@ -292,8 +293,9 @@ public class SpiderCodegen {
     append(");\n");
 
     append("void free_" + pg.getName() + "();\n");
-    append("std::map<lrtFct, PapifyConfig*> get_" + pg.getName() + "_papifyConfigs();\n");
-    append("void free_" + pg.getName() + "_papifyConfigs(std::map<lrtFct, PapifyConfig*>& map);\n");
+    append("std::map<lrtFct, std::map<const char *, PapifyConfig*>> get_" + pg.getName() + "_papifyConfigs();\n");
+    append("void free_" + pg.getName()
+        + "_papifyConfigs(std::map<lrtFct, std::map<const char *, PapifyConfig*>>& map);\n");
     append("\n");
 
     /* Core */
@@ -409,32 +411,36 @@ public class SpiderCodegen {
     final ArrayList<AbstractActor> papifiedActors = new ArrayList<>();
 
     for (final AbstractActor actor : this.functionMap.keySet()) {
-      PapifyConfigActor corePapifyConfigGroups = papifyConfigManager.getCorePapifyConfigGroupActor(actor.getName());
+      PapifyConfigActor corePapifyConfigGroups = papifyConfigManager
+          .getCorePapifyConfigGroupActor(actor.getVertexPath());
       if (corePapifyConfigGroups != null) {
         papifiedActors.add(actor);
-        if (!generatePapifyConfig(corePapifyConfigGroups, actor, uniqueEventSets, eventSetID)) {
+        if (!generatePapifyConfig(corePapifyConfigGroups, papifyConfigManager, actor, uniqueEventSets, eventSetID)) {
           eventSetID++;
         }
-
       }
     }
 
-    append("std::map<lrtFct, PapifyConfig*> get_" + pg.getName() + "_papifyConfigs() {\n");
-    append("\tstd::map<lrtFct, PapifyConfig*> map;\n");
+    append("std::map<lrtFct, std::map<const char *, PapifyConfig*>> get_" + pg.getName() + "_papifyConfigs() {\n");
+    append("\tstd::map<lrtFct, std::map<const char *, PapifyConfig*>> map;\n");
+    append("\tstd::map<const char *, PapifyConfig*> mapPapifyConfigs;\n");
     append("\t// Initializing the map\n");
     for (final AbstractActor actor : papifiedActors) {
-      append("\tmap.insert(std::make_pair(" + pg.getName() + "_fcts["
-          + SpiderNameGenerator.getFunctionName(actor).toUpperCase() + "_FCT" + "], create" + actor.getName()
-          + "PapifyConfig()));\n");
+      append("\tmapPapifyConfigs = create" + actor.getName() + "PapifyConfig();\n");
+      append("\tif(!mapPapifyConfigs.empty()) {\n");
+      append("\t\tmap.insert(std::make_pair(" + pg.getName() + "_fcts["
+          + SpiderNameGenerator.getFunctionName(actor).toUpperCase() + "_FCT" + "], mapPapifyConfigs));\n");
+      append("\t}\n");
     }
     append("\treturn map;\n");
     append("}\n\n");
 
-    append("void free_" + pg.getName() + "_papifyConfigs(std::map<lrtFct, PapifyConfig*>& map) {\n");
-    append("\tstd::map<lrtFct, PapifyConfig*>::iterator it;\n");
+    append("void free_" + pg.getName()
+        + "_papifyConfigs(std::map<lrtFct, std::map<const char *, PapifyConfig*>>& map) {\n");
+    append("\tstd::map<lrtFct, std::map<const char *, PapifyConfig*>>::iterator it;\n");
     append("\t// Freeing memory of the map \n");
     append("\tfor(it = map.begin(); it != map.end(); ++it) { \n");
-    append("\t\tdelete it->second;\n");
+    append("\t\tdelete it->second.begin()->second;\n");
     append("\t}\n");
     append("}\n");
     // Returns the final C++ code
@@ -454,78 +460,124 @@ public class SpiderCodegen {
    *          The current event set ID
    * @return true if the actor has the same event set as an existing one, false else
    */
-  private boolean generatePapifyConfig(final PapifyConfigActor corePapifyConfigGroups, final AbstractActor actor,
+  private boolean generatePapifyConfig(final PapifyConfigActor corePapifyConfigGroups,
+      PapifyConfigManager papifyConfigManager, final AbstractActor actor,
       final HashMap<ArrayList<String>, Integer> uniqueEventSets, final Integer eventSetID) {
     Map<String, Set<PapiEvent>> configInfo = corePapifyConfigGroups.getPAPIEvents();
-    // PapiComponent papiComponent = corePapifyConfigGroups.getPAPIComponent();
-    // Set<PapiEvent> papiEvents = corePapifyConfigGroups.getPAPIEvents();
-
-    Set<PapiEvent> includedEvents = new LinkedHashSet<>();
-    ArrayList<String> eventNames = new ArrayList<>();
 
     boolean eventMonitoring = false;
     boolean timingMonitoring = false;
 
     final PapiEvent timingEvent = new PapiEvent();
     timingEvent.setName("Timing");
-    /*
-     * for (PapiEvent event : papiEvents) { if (event.getName().equals(timingEvent.getName())) { timingMonitoring =
-     * true; } else if (papiComponent.containsEvent(event)) { includedEvents.add(event); eventMonitoring = true;
-     * eventNames.add(event.getName()); } }
-     */
-    // DM changed this to avoid errors
+
+    ArrayList<String> compNames = new ArrayList<>();
+    Map<String, ArrayList<String>> associatedEvents = new LinkedHashMap<>();
+
     for (String compName : configInfo.keySet()) {
+      // Build the eventNames and the Timing variables to be printed
       for (PapiEvent event : configInfo.get(compName)) {
         if (event.getName().equals(timingEvent.getName())) {
           timingMonitoring = true;
         } else {
-          includedEvents.add(event);
+          if (associatedEvents.get(compName) == null) {
+            ArrayList<String> compEventNames = new ArrayList<>();
+            compEventNames.add(event.getName());
+            associatedEvents.put(compName, compEventNames);
+          } else {
+            associatedEvents.get(compName).add(event.getName());
+          }
           eventMonitoring = true;
-          eventNames.add(event.getName());
         }
       }
+      // Build the peType variable to be printed
+      if (!compName.equals("Timing")) {
+        compNames.add(compName);
+      }
     }
+
     // Check if this set of ID already exists
     Integer realEventSetID = eventSetID;
     boolean found = false;
-    if (uniqueEventSets.isEmpty()) {
-      uniqueEventSets.put(eventNames, eventSetID);
-    } else {
-      for (Map.Entry<ArrayList<String>, Integer> eventSet : uniqueEventSets.entrySet()) {
-        final ArrayList<String> currentEventSetNames = eventSet.getKey();
-        final Integer currentEventSetID = eventSet.getValue();
-        if (eventNames.size() != currentEventSetNames.size()) {
-          continue;
-        } else {
-          found = eventNames.containsAll(currentEventSetNames) && currentEventSetNames.containsAll(eventNames);
-          if (found) {
-            realEventSetID = currentEventSetID;
-            break;
+    for (String compName : associatedEvents.keySet()) {
+      found = false;
+      final ArrayList<String> compEventSetNames = associatedEvents.get(compName);
+      if (!uniqueEventSets.isEmpty()) {
+        for (Map.Entry<ArrayList<String>, Integer> eventSet : uniqueEventSets.entrySet()) {
+          final ArrayList<String> currentEventSetNames = eventSet.getKey();
+          final Integer currentEventSetID = eventSet.getValue();
+          if (compEventSetNames.size() == currentEventSetNames.size()) {
+            found = compEventSetNames.containsAll(currentEventSetNames)
+                && currentEventSetNames.containsAll(compEventSetNames);
+            if (found) {
+              realEventSetID = currentEventSetID;
+              break;
+            }
           }
+        }
+        // If it does not already exist, add it to the set
+        if (!found) {
+          uniqueEventSets.put(compEventSetNames, eventSetID);
+        }
+      } else {
+        uniqueEventSets.put(compEventSetNames, eventSetID);
+      }
+    }
+
+    append("static std::map<const char *, PapifyConfig*> " + "create" + actor.getName() + "PapifyConfig() {\n");
+    append("\t// Setting the PapifyConfigs for actor: " + actor.getName() + "\n");
+    append("\tstd::map<const char *, PapifyConfig*> configMap;\n");
+    for (String compNameGen : compNames) {
+      append("\n\tPapifyConfig* config_" + compNameGen + "  = new PapifyConfig;\n");
+      append("\tconfig_" + compNameGen + "->peID_            = \"\";\n");
+      append("\tconfig_" + compNameGen + "->peType_          = \"" + compNameGen + "\";\n");
+      append("\tconfig_" + compNameGen + "->actorName_       = \"" + actor.getName() + "\";\n");
+      append("\tconfig_" + compNameGen + "->eventSize_       = "
+          + Integer.toString(associatedEvents.get(compNameGen).size()) + ";\n");
+      append("\tconfig_" + compNameGen + "->eventSetID_      = " + realEventSetID.toString() + ";\n");
+      final String timing = timingMonitoring ? "true" : "false";
+      append("\tconfig_" + compNameGen + "->isTiming_        = " + timing + ";\n");
+      if (eventMonitoring) {
+        append("\tconfig_" + compNameGen + "->monitoredEvents_ = std::vector<const char*>("
+            + Integer.toString(associatedEvents.get(compNameGen).size()) + ");\n");
+        int i = 0;
+        for (String name : associatedEvents.get(compNameGen)) {
+          append("\tconfig_" + compNameGen + "->monitoredEvents_[" + Integer.toString(i++) + "] = \"" + name + "\";\n");
         }
       }
     }
 
-    append("static PapifyConfig* " + "create" + actor.getName() + "PapifyConfig() {\n");
-    append("\tPapifyConfig* config  = new PapifyConfig;\n\n");
-    append("\t// Setting the PapifyConfig for actor: " + actor.getName() + "\n");
-    append("\tconfig->peID_            = 0;\n");
-    // append("\tconfig->peType_ = \"" + papiComponent.getId() + "\";\n");
-    // DM changed this to avoid errors
-    append("\tconfig->peType_          = \"" + configInfo.keySet() + "\";\n");
-    append("\tconfig->actorName_       = \"" + actor.getName() + "\";\n");
-    append("\tconfig->eventSize_       = " + Integer.toString(eventNames.size()) + ";\n");
-    append("\tconfig->eventSetID_      = " + realEventSetID.toString() + ";\n");
-    final String timing = timingMonitoring ? "true" : "false";
-    append("\tconfig->isTiming_        = " + timing + ";\n");
-    if (eventMonitoring) {
-      append("\tconfig->monitoredEvents_ = std::vector<const char*>(" + Integer.toString(eventNames.size()) + ");\n");
-      int i = 0;
-      for (String name : eventNames) {
-        append("\tconfig->monitoredEvents_[" + Integer.toString(i++) + "] = \"" + name + "\";\n");
+    if (timingMonitoring) {
+      append("\n\tPapifyConfig* config_Timing  = new PapifyConfig;\n");
+      append("\tconfig_Timing->peID_            = \"\";\n");
+      append("\tconfig_Timing->peType_          = \"\";\n");
+      append("\tconfig_Timing->actorName_       = \"" + actor.getName() + "\";\n");
+      append("\tconfig_Timing->eventSize_       = 0;\n");
+      append("\tconfig_Timing->eventSetID_      = " + realEventSetID.toString() + ";\n");
+      final String timing = timingMonitoring ? "true" : "false";
+      append("\tconfig_Timing->isTiming_        = " + timing + ";\n");
+    }
+
+    boolean configAssociated = false;
+    append("\n\t// Mapping actor to LRT PAPIFY configuration: " + actor.getName() + "\n");
+    for (String coreType : this.coresFromCoreType.keySet()) {
+      for (ComponentInstance compInst : this.coresFromCoreType.get(coreType)) {
+        configAssociated = false;
+        PapifyConfigPE configType = papifyConfigManager.getCorePapifyConfigGroupPE(coreType);
+        for (String compType : configType.getPAPIComponentIDs()) {
+          if (!compType.equals("Timing") && compNames.contains(compType)) {
+            configAssociated = true;
+            append("\tconfigMap.insert(std::make_pair(\"LRT_" + this.coreIds.get(compInst.getInstanceName())
+                + "\", config_" + compType + "));\n");
+          }
+        }
+        if (!configAssociated && timingMonitoring) {
+          append("\tconfigMap.insert(std::make_pair(\"LRT_" + this.coreIds.get(compInst.getInstanceName())
+              + "\", config_Timing));\n");
+        }
       }
     }
-    append("\treturn config;\n");
+    append("\treturn configMap;\n");
     append("}\n\n");
     return found;
   }